+++ packages/docs/components/atoms/Action/stories/prevent-default/app.js
@@ -1,0 +1,20 @@
+import { Base, createApp } from '@studiometa/js-toolkit';
+import { Action } from '@studiometa/ui';
+
+class Foo extends Base {
+  static config = {
+    name: 'Foo',
+  };
+}
+
+class App extends Base {
+  static config = {
+    name: 'App',
+    components: {
+      Action,
+      Foo,
+    },
+  };
+}
+
+export default createApp(App);