+++ packages/tests/__utils__/index.ts
@@ -1,1 +1,4 @@
 export * from './h.js';
+export * from './lifecycle.js';
+export * from './mockIntersectionObserver.js';
+export * from './wait.js';