+++ packages/tests/__utils__/wait.ts
@@ -1,3 +1,3 @@
-export default function wait(delay = 100) {
+export function wait(delay = 100) {
   return new Promise(resolve => setTimeout(resolve, delay));
 }