+++ packages/ui/atoms/Action/Action.ts
@@ -1,4 +1,4 @@
-import { Base } from '@studiometa/js-toolkit';
+import { Base, getInstances } from '@studiometa/js-toolkit';
 import { isFunction } from '@studiometa/js-toolkit/utils';
 import type { BaseProps, BaseConfig } from '@studiometa/js-toolkit';
 
@@ -6,10 +6,18 @@ export interface ActionProps extends BaseProps {
   $options: {
     on: string;
     target: string;
-    method: string;
     selector: string;
+    effect: string;
   };
 }
+
+/**
+ * Extract component name and an optional additional selector from a string.
+ * @type {RegExp}
+ */
+const TARGET_REGEX = /([a-zA-Z]+)(\((.*)\))?/;
+
+const effectCache = new Map<string, Function>();
 
 /**
  * Action class.
@@ -23,60 +31,90 @@ export class Action<T extends BaseProps = BaseProps> extends Base<ActionProps & 
         default: 'click',
       },
       target: String,
-      method: String,
       selector: String,
+      effect: String,
     },
   };
+
+  get event() {
+    const [event] = this.$options.on.split('.', 1);
+    return event;
+  }
+
+  get modifiers() {
+    return this.$options.on.split('.').slice(1);
+  }
+
+  get effect() {
+    const { effect } = this.$options;
+    if (!effectCache.has(effect)) {
+      effectCache.set(effect, new Function('ctx', 'event', 'target', `return ${effect}`));
+    }
+    return effectCache.get(effect);
+  }
+
+  get targets() {
+    const { target } = this.$options;
+    const parts = target.split(' ').map((part) => {
+      const [, name, , selector] = part.match(TARGET_REGEX) ?? [];
+      return [name, selector];
+    });
+
+    const targets = [] as Array<Record<string, Base>>;
+
+    for (const instance of getInstances()) {
+      const { name } = instance.__config;
+
+      for (const part of parts) {
+        const shouldPush =
+          part[0] === name && (!part[1] || (part[1] && instance.$el.matches(part[1])));
+        if (shouldPush) {
+          targets.push({ [instance.$options.name]: instance });
+        }
+      }
+    }
+
+    return targets;
+  }
 
   /**
    * Run method on targeted components
    */
-  handleEvent() {
-    const { target: componentNames, method, selector } = this.$options;
+  handleEvent(event: Event) {
+    const { targets, effect, modifiers } = this;
 
-    let targets = componentNames.includes(' ')
-      ? componentNames
-          .split(' ')
-          .flatMap((componentName) => this.$root.$children?.[componentName] as Base[])
-      : (this.$root.$children?.[componentNames] as Base[]);
-
-    if (!targets || !targets.length) {
-      this.$warn('Target not found.');
-      return;
+    if (modifiers.includes('prevent')) {
+      event.preventDefault();
     }
 
-    if (selector || selector.length > 0) {
-      targets = targets.filter((target) => target.$el.matches(selector));
+    if (modifiers.includes('stop')) {
+      event.stopPropagation();
     }
 
-    if (!targets || !targets.length) {
-      this.$warn(`Target with selector "${selector}" not found.`);
-      return;
-    }
-
-    targets.forEach((target) => {
-      if (!isFunction(target[method])) {
-        this.$warn(`Method "${method}()" not found on target.`, target);
-        return;
+    for (const target of targets) {
+      try {
+        const [name, currentTarget] = Object.entries(target).flat();
+        const value = effect(target, event, currentTarget);
+        if (typeof value === 'function') {
+          value(target, event, currentTarget);
+        }
+      } catch (err) {
+        this.$warn(err);
       }
-
-      target[method]();
-    });
+    }
   }
 
   /**
    * Mounted
    */
   mounted() {
-    const { on: eventName, target, method } = this.$options;
+    const { event, modifiers } = this;
 
-    if (!target || !method || target.length <= 0 || method.length <= 0) {
-      this.$warn('No target or method specified.');
-      this.$terminate();
-      return;
-    }
-
-    this.$el.addEventListener(eventName, this);
+    this.$el.addEventListener(event, this, {
+      capture: modifiers.includes('capture'),
+      once: modifiers.includes('once'),
+      passive: modifiers.includes('passive'),
+    });
   }
 
   /**